+++ spec/models/g5_authenticatable/user_spec.rb
@@ -29,11 +29,260 @@ describe G5Authenticatable::User do
     expect(user).to respond_to(:update_tracked_fields!)
   end
 
-  it { should validate_presence_of(:email) }
-  it { should validate_uniqueness_of(:email) }
-  it { should validate_uniqueness_of(:uid).scoped_to(:provider) }
+  it { is_expected.to validate_presence_of(:email) }
+  it { is_expected.to validate_uniqueness_of(:email) }
+  it { is_expected.to validate_uniqueness_of(:uid).scoped_to(:provider) }
 
   it 'should support timeouts' do
     expect(user.timeout_in).to be > 0
+  end
+
+  it 'should expose the first_name' do
+    expect(user.first_name).to eq(user_attributes[:first_name])
+  end
+
+  it 'should expose the last_name' do
+    expect(user.last_name).to eq(user_attributes[:last_name])
+  end
+
+  it 'should expose the phone number' do
+    expect(user.phone_number).to eq(user_attributes[:phone_number])
+  end
+
+  it 'should expose the title' do
+    expect(user.title).to eq(user_attributes[:title])
+  end
+
+  it 'should expose the organization name' do
+    expect(user.organization_name).to eq(user_attributes[:organization_name])
+  end
+
+  describe '.new_with_session' do
+    subject(:new_user) { G5Authenticatable::User.new_with_session(params, session) }
+
+    let(:params) { Hash.new }
+    let(:auth_data) do
+      OmniAuth::AuthHash.new({
+        'provider' => new_user_attributes[:provider],
+        'uid' => new_user_attributes[:uid],
+        'info' => {
+          'email' => new_user_attributes[:email],
+          'name' => "#{new_user_attributes[:first_name]} #{new_user_attributes[:last_name]}",
+          'first_name' => new_user_attributes[:first_name],
+          'last_name' => new_user_attributes[:last_name],
+          'phone' => new_user_attributes[:phone_number]
+        },
+        'credentials' => {
+          'token' => new_user_attributes[:g5_access_token],
+          'expires' => true,
+          'expires_at' => Time.now + 1000
+        },
+        'extra' => {
+          'title' => new_user_attributes[:title],
+          'organization_name' => new_user_attributes[:organization_name],
+          'raw_info' => {}
+        }
+      })
+    end
+    let(:new_user_attributes) { FactoryGirl.attributes_for(:g5_authenticatable_user) }
+
+    context 'when there is auth data in the session' do
+      let(:session) { {'omniauth.auth' => auth_data} }
+
+      it 'should initialize a new user' do
+        expect(new_user).to be_a_new_record
+      end
+
+      it 'should not persist the user' do
+        expect { new_user }.to_not change { G5Authenticatable::User.count }
+      end
+
+      it 'should set the provider from the session data' do
+        expect(new_user.provider).to eq(new_user_attributes[:provider])
+      end
+
+      it 'should set the uid from the session data' do
+        expect(new_user.uid).to eq(new_user_attributes[:uid])
+      end
+
+      it 'should set the email from the session data' do
+        expect(new_user.email).to eq(new_user_attributes[:email])
+      end
+
+      it 'should set the first_name from the session data' do
+        expect(new_user.first_name).to eq(new_user_attributes[:first_name])
+      end
+
+      it 'should set the last_name from the session data' do
+        expect(new_user.last_name).to eq(new_user_attributes[:last_name])
+      end
+
+      it 'should set the phone_number from the session data' do
+        expect(new_user.phone_number).to eq(new_user_attributes[:phone_number])
+      end
+
+      it 'should set the title from the session data' do
+        expect(new_user.title).to eq(new_user_attributes[:title])
+      end
+
+      it 'should set the organization_name from the session data' do
+        expect(new_user.organization_name).to eq(new_user_attributes[:organization_name])
+      end
+    end
+
+    context 'when there is no auth data in the session' do
+      let(:session) { Hash.new }
+
+      it 'should initialize a new user' do
+        expect(new_user).to be_a_new_record
+      end
+
+      it 'should not persist the user' do
+        expect { new_user }.to_not change { G5Authenticatable::User.count }
+      end
+
+      it 'should not set the uid' do
+        expect(new_user.uid).to be_nil
+      end
+
+      it 'should set the provider to the default value' do
+        expect(new_user.provider).to eq('g5')
+      end
+
+      it 'should not set the email' do
+        expect(new_user.email).to be_blank
+      end
+    end
+  end
+
+  describe '.find_and_update_for_g5_oauth' do
+    subject(:updated_user) { G5Authenticatable::User.find_and_update_for_g5_oauth(auth_data) }
+
+    let(:user_attributes) do
+      FactoryGirl.attributes_for(:g5_authenticatable_user,
+                                 first_name: nil,
+                                 last_name: nil,
+                                 phone_number: nil,
+                                 title: nil,
+                                 organization_name: nil
+                                )
+    end
+    before { user }
+
+    let(:auth_data) do
+      OmniAuth::AuthHash.new({
+        'provider' => user_attributes[:provider],
+        'uid' => user_attributes[:uid],
+        'info' => {
+          'email' => updated_attributes[:email],
+          'first_name' => updated_attributes[:first_name],
+          'last_name' => updated_attributes[:last_name],
+          'phone' => updated_attributes[:phone_number]
+        },
+        'credentials' => {
+          'token' => updated_attributes[:g5_access_token],
+          'expires' => true,
+          'expires_at' => Time.now + 1000
+        },
+        'extra' => {
+          'title' => updated_attributes[:title],
+          'organization_name' => updated_attributes[:organization_name],
+          'raw_info' => {}
+        }
+      })
+    end
+
+    context 'when user info is the same' do
+      let(:updated_attributes) do
+        user_attributes.merge(g5_access_token: 'updatedtoken42')
+      end
+
+      it 'should update the access token' do
+        expect { updated_user }.to change { user.reload.g5_access_token }.to(updated_attributes[:g5_access_token])
+      end
+
+      it 'should return the updated user' do
+        expect(updated_user).to eq(user.reload)
+      end
+
+      it 'should not change the user email' do
+        expect { updated_user }.to_not change { user.reload.email }
+      end
+
+      it 'should not change the user first_name' do
+        expect { updated_user }.to_not change { user.reload.first_name }
+      end
+
+      it 'should not change the user last_name' do
+        expect { updated_user }.to_not change { user.reload.last_name }
+      end
+
+      it 'should not change the user phone_number' do
+        expect { updated_user }.to_not change { user.reload.phone_number }
+      end
+
+      it 'should not change the user title' do
+        expect { updated_user }.to_not change { user.reload.title }
+      end
+
+      it 'should not change the user organization_name' do
+        expect { updated_user }.to_not change { user.reload.organization_name }
+      end
+    end
+
+    context 'when user info has changed' do
+      let(:updated_attributes) do
+        {
+          uid: user.uid,
+          provider: user.provider,
+          g5_access_token: 'updatedtoken42',
+          first_name: 'Updated First Name',
+          last_name: 'Updated Last Name',
+          phone_number: '555.555.5555 x123',
+          title: 'Recently Promoted',
+          organization_name: 'Updated Department'
+        }
+      end
+
+      it 'should update the access token' do
+        expect { updated_user }.to change { user.reload.g5_access_token }.to(updated_attributes[:g5_access_token])
+      end
+
+      it 'should return the updated user' do
+        expect(updated_user).to eq(user.reload)
+      end
+
+      it 'should not change the uid' do
+        expect { updated_user }.to_not change { user.reload.uid }
+      end
+
+      it 'should not change the provider' do
+        expect { updated_user }.to_not change { user.reload.provider }
+      end
+
+      it 'should not change the email' do
+        expect { updated_user }.to_not change { user.reload.email }
+      end
+
+      it 'should update the first name' do
+        expect { updated_user }.to change { user.reload.first_name }.to(updated_attributes[:first_name])
+      end
+
+      it 'should update the last name' do
+        expect { updated_user }.to change { user.reload.last_name }.to(updated_attributes[:last_name])
+      end
+
+      it 'should update the phone number' do
+        expect { updated_user }.to change { user.reload.phone_number }.to(updated_attributes[:phone_number])
+      end
+
+      it 'should update the title' do
+        expect { updated_user }.to change { user.reload.title }.to(updated_attributes[:title])
+      end
+
+      it 'should update the organization_name' do
+        expect { updated_user }.to change { user.reload.organization_name }.to(updated_attributes[:organization_name])
+      end
+    end
   end
 end